+++ bytebank/src/main/kotlin/br/com/alura/bytebank/modelo/ContaCorrente.kt
@@ -1,4 +1,4 @@
-package modelo
+package br.com.alura.bytebank.modelo
 
 class ContaCorrente(
     titular: String,
