+++ bytebank/src/main/kotlin/br/com/alura/bytebank/modelo/Diretor.kt
@@ -1,10 +1,11 @@
-package modelo
+package br.com.alura.bytebank.modelo
 
-class Gerente(
+class Diretor(
     nome: String,
     cpf: String,
     salario: Double,
-    senha: Int
+    senha: Int,
+    val plr: Double
 ) : FuncionarioAdmin(
     nome = nome,
     cpf = cpf,
@@ -14,7 +15,7 @@ class Gerente(
 
     override val bonificacao: Double
         get() {
-            println("Bonificação modelo.Gerente")
-            return salario
+            println("Bonificação br.com.alura.bytebank.modelo.Diretor")
+            return salario + plr
         }
 }
