+++ bytebank/src/main/kotlin/br/com/alura/bytebank/modelo/FuncionarioAdmin.kt
@@ -1,4 +1,4 @@
-package modelo
+package br.com.alura.bytebank.modelo
 
 abstract class FuncionarioAdmin(
     nome: String,
