+++ bytebank/src/main/kotlin/br/com/alura/bytebank/teste/testaComportamentosConta.kt
@@ -1,5 +1,5 @@
-import modelo.ContaCorrente
-import modelo.ContaPoupanca
+import br.com.alura.bytebank.modelo.ContaCorrente
+import br.com.alura.bytebank.modelo.ContaPoupanca
 
 fun testaComportamentosConta() {
     val contaVinicius = ContaCorrente(titular = "Vinicius", numero = 1000)
