+++ bytebank/src/main/kotlin/br/com/alura/bytebank/teste/testaCopiasEReferencias.kt
@@ -1,5 +1,5 @@
-import modelo.ContaCorrente
-import modelo.ContaPoupanca
+import br.com.alura.bytebank.modelo.ContaCorrente
+import br.com.alura.bytebank.modelo.ContaPoupanca
 
 fun testaCopiasEReferencias() {
     val numeroX = 10
@@ -14,8 +14,8 @@ fun testaCopiasEReferencias() {
     val contaJoao = ContaCorrente("João", 1002)
     var contaMaria = ContaPoupanca("Maria", 1003)
 
-    println("Titular modelo.Conta João: ${contaJoao.titular}")
-    println("Titular modelo.Conta Maria: ${contaMaria.titular}")
+    println("Titular br.com.alura.bytebank.modelo.Conta João: ${contaJoao.titular}")
+    println("Titular br.com.alura.bytebank.modelo.Conta Maria: ${contaMaria.titular}")
 
     println(contaJoao)
     println(contaMaria)
